Fix signed casts causing build-specific behavior in random names
Random letter choices in make_name were casting a pseudo-random unsigned
32bit integer to int and then modding it with something of type size_t,
which will usually implicitly cast it back to size_t (which is unsigned
and may vary across platforms; the cast is triggered because it
typically is bigger than the int type so has a higher rank). If the int
value ended up as negative in the first cast, the implicit cast to
size_t was coming out differently on different platforms, leading to
different random names (basically depending on whether it was a 32bit or
64bit build).  This change maintains unsignedness throughout for more
stable behavior.

Because random naming uses its own rng objects, the effect of this was
partitioned off from anything else in levelgen, leading to identical
properties on randarts with different names across platforms in gotm 2
(since windows is a 32bit build).

+++ crawl-ref/source/item-name.cc
@@ -61,9 +61,9 @@
 #include "viewgeom.h"
 
 static bool _is_consonant(char let);
-static char _random_vowel(int seed);
-static char _random_cons(int seed);
-static string _random_consonant_set(int seed);
+static char _random_vowel(uint32_t seed);
+static char _random_cons(uint32_t seed);
+static string _random_consonant_set(size_t seed);
 
 static void _maybe_identify_pack_item()
 {
@@ -2766,7 +2766,6 @@ string make_name(uint32_t seed, makename_type name_type)
                                               : '\0';
         const char penult_char = name.length() > 1 ? name[name.length() - 2]
                                                     : '\0';
-
         if (name.empty() && name_type == MNAME_JIYVA)
         {
             // Start the name with a predefined letter.
@@ -2923,7 +2922,7 @@ static bool _is_consonant(char let)
 
 // Returns a random vowel (a, e, i, o, u with equal probability) or space
 // or 'y' with lower chances.
-static char _random_vowel(int seed)
+static char _random_vowel(uint32_t seed)
 {
     static const char vowels[] = "aeiouaeiouaeiouy  ";
     return vowels[ seed % (sizeof(vowels) - 1) ];
@@ -2931,7 +2930,7 @@ static char _random_vowel(int seed)
 
 // Returns a random consonant with not quite equal probability.
 // Does not include 'y'.
-static char _random_cons(int seed)
+static char _random_cons(uint32_t seed)
 {
     static const char consonants[] = "bcdfghjklmnpqrstvwxzcdfghlmnrstlmnrst";
     return consonants[ seed % (sizeof(consonants) - 1) ];
@@ -2946,7 +2945,7 @@ static char _random_cons(int seed)
  * @return      A random length 2 or 3 consonant set; e.g. "kl", "str", etc.
  *              If the seed is out of bounds, return "";
  */
-static string _random_consonant_set(int seed)
+static string _random_consonant_set(size_t seed)
 {
     // Pick a random combination of consonants from the set below.
     //   begin  -> [RCS_BB, RCS_EB) = [ 0, 27)
@@ -2976,7 +2975,7 @@ static string _random_consonant_set(int seed)
     };
     COMPILE_CHECK(ARRAYSZ(consonant_sets) == RCS_END);
 
-    ASSERT_RANGE(seed, 0, (int) ARRAYSZ(consonant_sets));
+    ASSERT_RANGE(seed, 0, ARRAYSZ(consonant_sets));
 
     return consonant_sets[seed];
 }
